Copter: reset yaw mode to default when yaw and yaw_rate ignored
Currently, when the functions 'set_destination', 'set_velocity' and 'set_destination_posvel' are called with the use_yaw or use_yaw_rate flags to true,
the auto_yaw mode is set to fixed_yaw or rate repectively. But if a second call of the same function comes with both flags to false, the auto_yaw mode
remains in the previous mode, rather than using default yaw mode. This is a problem particularly when the user wants to move at a fixed yaw and then
switch back to face the target default mode (defined by WP_YAW_BEHAVIOR).

The fix for this is to add 'auto_yaw.set_mode_to_default' in the else case of the function 'set_yaw_state'. This implementation allows the user to
switch back to default yaw mode defined by WP_YAW_BEHAVIOR when sending a new SET_POSITION_TARGET_LOCAL_NED or SET_POSITION_TARGET_GLOBAL_INT with
the mask ignoring both yaw and yaw rate.